refact(Invoices): reorganize redux files
