+++ src/components/invoices/show/header.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Link } from 'react-router-dom';
 
-import { Invoice } from 'model-types';
+import { Invoice } from 'services/invoices';
 import { formatDateRange } from 'utils/date';
 
 export interface Props {
