+++ src/components/invoices/show/items.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { Invoice } from 'model-types';
+import { Invoice } from 'services/invoices';
 import { formatMoney } from 'utils/money';
 import { formatDate } from 'utils/date';
