+++ src/reducers/index.ts
@@ -1,8 +1,10 @@
 export { default as app }                  from './app';
 export { default as auth }                 from 'services/auth/reducer.js';
 export { default as flashMessages }        from 'services/flash-messages/reducer.js';
-export { default as invoice }              from './invoice';
-export { default as invoices }             from './invoices';
+export { default as invoice }              from 'services/invoices/entity-reducer.js';
+export { default as invoices }             from 'services/invoices/entities-reducer.js';
+export { default as unpaidInvoices }       from 'services/invoices/unpaid-reducer.js';
+export { default as unpaidInvoicesCount }  from 'services/invoices/unpaid-count-reducer.js';
 export { default as organizations }        from './organizations';
 export { default as currentOrganization }  from './current-organization';
 export { default as transaction }          from './transaction';
@@ -15,5 +17,3 @@ export { default as category }             from './category';
 export { default as customers }            from 'services/customers/entities-reducer.js';
 export { default as customer }             from 'services/customers/entity-reducer.js';
 export { default as balances }             from './balances';
-export { default as unpaidInvoices }       from './unpaid-invoices';
-export { default as unpaidInvoicesCount }  from './unpaid-invoices-count';
